Disentangle Store and Component functionality
The Store() interface provides write-acess to the component storage. As
such, the Component class was delegating a lot of logic like capturing
output, status writes etc. to the Store. This commit removes such
Component-related concerns from the Store and moves it into the
Component. In future, the Store() write interface might be completely
replaced by a StorageFilsystemModel
