+++ src/machinable/config/parser.py
@@ -2,12 +2,12 @@ import importlib
 import inspect
 
 import regex
+from flatten_dict import unflatten
 
 from ..utils.dicts import get_or_fail, read_path_dict, update_dict
 from ..utils.formatting import exception_to_str, msg
 from ..utils.utils import is_valid_variable_name
 from .mapping import _reserved_keys, _used_keys
-from flatten_dict import unflatten
 
 
 class ModuleClass(object):
