+++ tests/storage/storage_test.py
@@ -1,4 +1,5 @@
 import os
+
 import pytest
 
 from machinable import Storage
